refactor: code-quality cleanups (ACK helpers, generated arity, build hygiene)
Protocol
- Add sys_error()/internal_error() in commands/utils.rs and convert the 63
  ACK 52 call sites whose wording was exactly "Error: {e}" or "internal
  error". Sites with their own wording were deliberately left alone rather
  than normalized, so the emitted bytes are unchanged everywhere.
- Generate command_arity() from the #[command(...)] attribute via a new
  `args = "MIN..MAX"` key instead of hand-listing all 136 command names a
  second time. Verified mechanically: all 113 tabulated (min,max) pairs are
  identical to the pre-refactor table, and the five untabulated names still
  return None.
- commands/notcommands omit getfingerprint when built without the
  fingerprint feature, so clients aren't offered a command that only ACKs.

Build
- chromaprint-sys-next (a C binding) is now behind a default-on
  `fingerprint` feature: `--no-default-features` drops the C dependency
  entirely, while the normal build is unchanged and still fingerprints.
- [profile.dev.package."*"] opt-level = 2 so the DSP-heavy dependency graph
  isn't run at -O0 under cargo test.
- Drop redundant_field_names, unnecessary_cast and same_item_push from the
  workspace clippy allow-list and fix the three real sites they exposed in
  rmpd-player. Both the default and --no-default-features builds are
  warning-free.
- Document the duplicate base64/rand/thiserror majors in deny.toml: they
  come from third-party crates we don't control, so the multiple-versions
  check stays actionable instead of permanently noisy.
- CI.md: correct the now-stale clippy allow-list.

+++ rmpd-player/src/engine.rs
@@ -703,7 +703,7 @@ impl PlaybackEngine {
                 };
                 if let Some(end_samples) = cf_end_samples {
                     // Sample offset at which the overlap window begins
-                    let cf_window = crossfade_secs as u64 * samples_per_second as u64;
+                    let cf_window = crossfade_secs as u64 * samples_per_second;
                     let cf_start = end_samples.saturating_sub(cf_window);
                     // mpd refuses to cross-fade a track too short to fit the
                     // window (mirrors `CrossFadeSettings::CanCrossFadeSong`,
